+++ src/kernel/bstruct/givarrayfixed.h
@@ -35,7 +35,7 @@ namespace Givaro {
     {
         T  _data[SIZE];        // _data
     public :
-        typedef int  			Indice_t;
+        typedef size_t			Indice_t;
         typedef T    			Type_t;
         typedef ArrayFixed<T,SIZE> 	Self_t;
         typedef Type_t		*Iterator_t;
